LayoutTests/imported/w3c/web-platform-tests/streams/idlharness.any-expected.txt


PASS idl_test setup 
PASS idl_test validation 
PASS ReadableStream interface: existence and properties of interface object 
PASS ReadableStream interface object length 
PASS ReadableStream interface object name 
PASS ReadableStream interface: existence and properties of interface prototype object 
PASS ReadableStream interface: existence and properties of interface prototype object's "constructor" property 
PASS ReadableStream interface: existence and properties of interface prototype object's @@unscopables property 
FAIL ReadableStream interface: attribute locked assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation cancel(optional any) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation getReader(optional ReadableStreamGetReaderOptions) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation pipeThrough(ReadableWritablePair, optional StreamPipeOptions) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation pipeTo(WritableStream, optional StreamPipeOptions) assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: operation tee() assert_true: property should be enumerable expected true got false
FAIL ReadableStream interface: async iterable<any> undefined is not an object (evaluating 'iteratorDesc.writable')
PASS ReadableStream must be primary interface of new ReadableStream() 
PASS Stringification of new ReadableStream() 
PASS ReadableStream interface: new ReadableStream() must inherit property "locked" with the proper type 
PASS ReadableStream interface: new ReadableStream() must inherit property "cancel(optional any)" with the proper type 
PASS ReadableStream interface: calling cancel(optional any) on new ReadableStream() with too few arguments must throw TypeError 
PASS ReadableStream interface: new ReadableStream() must inherit property "getReader(optional ReadableStreamGetReaderOptions)" with the proper type 
PASS ReadableStream interface: calling getReader(optional ReadableStreamGetReaderOptions) on new ReadableStream() with too few arguments must throw TypeError 
PASS ReadableStream interface: new ReadableStream() must inherit property "pipeThrough(ReadableWritablePair, optional StreamPipeOptions)" with the proper type 
PASS ReadableStream interface: calling pipeThrough(ReadableWritablePair, optional StreamPipeOptions) on new ReadableStream() with too few arguments must throw TypeError 
PASS ReadableStream interface: new ReadableStream() must inherit property "pipeTo(WritableStream, optional StreamPipeOptions)" with the proper type 
FAIL ReadableStream interface: calling pipeTo(WritableStream, optional StreamPipeOptions) on new ReadableStream() with too few arguments must throw TypeError assert_unreached: Throws "TypeError: undefined is not an object" instead of rejecting promise Reached unreachable code
PASS ReadableStream interface: new ReadableStream() must inherit property "tee()" with the proper type 
FAIL ReadableStreamDefaultReader interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface object length assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface object name assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: attribute closed assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: operation cancel(optional any) assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: operation read() assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader interface: operation releaseLock() assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
FAIL ReadableStreamDefaultReader must be primary interface of (new ReadableStream()).getReader() assert_own_property: self does not have own property "ReadableStreamDefaultReader" expected property "ReadableStreamDefaultReader" missing
PASS Stringification of (new ReadableStream()).getReader() 
PASS ReadableStreamDefaultReader interface: (new ReadableStream()).getReader() must inherit property "closed" with the proper type 
PASS ReadableStreamDefaultReader interface: (new ReadableStream()).getReader() must inherit property "cancel(optional any)" with the proper type 
PASS ReadableStreamDefaultReader interface: calling cancel(optional any) on (new ReadableStream()).getReader() with too few arguments must throw TypeError 
PASS ReadableStreamDefaultReader interface: (new ReadableStream()).getReader() must inherit property "read()" with the proper type 
PASS ReadableStreamDefaultReader interface: (new ReadableStream()).getReader() must inherit property "releaseLock()" with the proper type 
FAIL ReadableStreamBYOBReader interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface object length assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface object name assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: attribute closed assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: operation cancel(optional any) assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: operation read(ArrayBufferView) assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader interface: operation releaseLock() assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
FAIL ReadableStreamBYOBReader must be primary interface of (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) assert_own_property: self does not have own property "ReadableStreamBYOBReader" expected property "ReadableStreamBYOBReader" missing
PASS Stringification of (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) 
PASS ReadableStreamBYOBReader interface: (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) must inherit property "closed" with the proper type 
PASS ReadableStreamBYOBReader interface: (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) must inherit property "cancel(optional any)" with the proper type 
PASS ReadableStreamBYOBReader interface: calling cancel(optional any) on (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) with too few arguments must throw TypeError 
PASS ReadableStreamBYOBReader interface: (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) must inherit property "read(ArrayBufferView)" with the proper type 
PASS ReadableStreamBYOBReader interface: calling read(ArrayBufferView) on (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) with too few arguments must throw TypeError 
PASS ReadableStreamBYOBReader interface: (new ReadableStream({ type: 'bytes' })).getReader({ mode: 'byob' }) must inherit property "releaseLock()" with the proper type 
FAIL ReadableStreamDefaultController interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface object length assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface object name assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: attribute desiredSize assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: operation close() assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: operation enqueue(optional any) assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController interface: operation error(optional any) assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
FAIL ReadableStreamDefaultController must be primary interface of self.readableStreamDefaultController assert_own_property: self does not have own property "ReadableStreamDefaultController" expected property "ReadableStreamDefaultController" missing
PASS Stringification of self.readableStreamDefaultController 
PASS ReadableStreamDefaultController interface: self.readableStreamDefaultController must inherit property "desiredSize" with the proper type 
PASS ReadableStreamDefaultController interface: self.readableStreamDefaultController must inherit property "close()" with the proper type 
PASS ReadableStreamDefaultController interface: self.readableStreamDefaultController must inherit property "enqueue(optional any)" with the proper type 
PASS ReadableStreamDefaultController interface: calling enqueue(optional any) on self.readableStreamDefaultController with too few arguments must throw TypeError 
PASS ReadableStreamDefaultController interface: self.readableStreamDefaultController must inherit property "error(optional any)" with the proper type 
PASS ReadableStreamDefaultController interface: calling error(optional any) on self.readableStreamDefaultController with too few arguments must throw TypeError 
FAIL ReadableByteStreamController interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface object length assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface object name assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: attribute byobRequest assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: attribute desiredSize assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: operation close() assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: operation enqueue(ArrayBufferView) assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController interface: operation error(optional any) assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
FAIL ReadableByteStreamController must be primary interface of self.readableByteStreamController assert_own_property: self does not have own property "ReadableByteStreamController" expected property "ReadableByteStreamController" missing
PASS Stringification of self.readableByteStreamController 
FAIL ReadableByteStreamController interface: self.readableByteStreamController must inherit property "byobRequest" with the proper type assert_in_array: wrong type: not object or function value "undefined" not in array ["object", "function"]
PASS ReadableByteStreamController interface: self.readableByteStreamController must inherit property "desiredSize" with the proper type 
PASS ReadableByteStreamController interface: self.readableByteStreamController must inherit property "close()" with the proper type 
PASS ReadableByteStreamController interface: self.readableByteStreamController must inherit property "enqueue(ArrayBufferView)" with the proper type 
PASS ReadableByteStreamController interface: calling enqueue(ArrayBufferView) on self.readableByteStreamController with too few arguments must throw TypeError 
PASS ReadableByteStreamController interface: self.readableByteStreamController must inherit property "error(optional any)" with the proper type 
PASS ReadableByteStreamController interface: calling error(optional any) on self.readableByteStreamController with too few arguments must throw TypeError 
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface object assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface object length assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface object name assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface prototype object assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: attribute view assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: operation respond(unsigned long long) assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest interface: operation respondWithNewView(ArrayBufferView) assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
FAIL ReadableStreamBYOBRequest must be primary interface of self.readableStreamByobRequest assert_own_property: self does not have own property "ReadableStreamBYOBRequest" expected property "ReadableStreamBYOBRequest" missing
PASS Stringification of self.readableStreamByobRequest 
PASS ReadableStreamBYOBRequest interface: self.readableStreamByobRequest must inherit property "view" with the proper type 
PASS ReadableStreamBYOBRequest interface: self.readableStreamByobRequest must inherit property "respond(unsigned long long)" with the proper type 
FAIL ReadableStreamBYOBRequest interface: calling respond(unsigned long long) on self.readableStreamByobRequest with too few arguments must throw TypeError assert_throws_js: Called with 0 arguments function "function () {
            fn.apply(obj, args);
        }" threw object "RangeError: bytesWritten has an incorrect value" ("RangeError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
PASS ReadableStreamBYOBRequest interface: self.readableStreamByobRequest must inherit property "respondWithNewView(ArrayBufferView)" with the proper type 
PASS ReadableStreamBYOBRequest interface: calling respondWithNewView(ArrayBufferView) on self.readableStreamByobRequest with too few arguments must throw TypeError 
PASS WritableStream interface: existence and properties of interface object 
PASS WritableStream interface object length 
PASS WritableStream interface object name 
PASS WritableStream interface: existence and properties of interface prototype object 
PASS WritableStream interface: existence and properties of interface prototype object's "constructor" property 
PASS WritableStream interface: existence and properties of interface prototype object's @@unscopables property 
FAIL WritableStream interface: attribute locked assert_true: The prototype object must have a property "locked" expected true got false
FAIL WritableStream interface: operation abort(optional any) assert_equals: property has wrong .length expected 0 but got 1
PASS WritableStream interface: operation close() 
FAIL WritableStream interface: operation getWriter() assert_own_property: interface prototype object missing non-static operation expected property "getWriter" missing
PASS WritableStream must be primary interface of new WritableStream() 
PASS Stringification of new WritableStream() 
FAIL WritableStream interface: new WritableStream() must inherit property "locked" with the proper type assert_inherits: property "locked" not found in prototype chain
PASS WritableStream interface: new WritableStream() must inherit property "abort(optional any)" with the proper type 
PASS WritableStream interface: calling abort(optional any) on new WritableStream() with too few arguments must throw TypeError 
PASS WritableStream interface: new WritableStream() must inherit property "close()" with the proper type 
FAIL WritableStream interface: new WritableStream() must inherit property "getWriter()" with the proper type assert_inherits: property "getWriter" not found in prototype chain
FAIL WritableStreamDefaultWriter interface: existence and properties of interface object assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface object length assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface object name assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: existence and properties of interface prototype object assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: attribute closed assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: attribute desiredSize assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: attribute ready assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: operation abort(optional any) assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: operation close() assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: operation releaseLock() assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter interface: operation write(optional any) assert_own_property: self does not have own property "WritableStreamDefaultWriter" expected property "WritableStreamDefaultWriter" missing
FAIL WritableStreamDefaultWriter must be primary interface of (new WritableStream()).getWriter() assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL Stringification of (new WritableStream()).getWriter() assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "closed" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "desiredSize" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "ready" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "abort(optional any)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: calling abort(optional any) on (new WritableStream()).getWriter() with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "close()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "releaseLock()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: (new WritableStream()).getWriter() must inherit property "write(optional any)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultWriter interface: calling write(optional any) on (new WritableStream()).getWriter() with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: (new WritableStream()).getWriter is not a function. (In '(new WritableStream()).getWriter()', '(new WritableStream()).getWriter' is undefined)"
FAIL WritableStreamDefaultController interface: existence and properties of interface object assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController interface object length assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController interface object name assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController interface: existence and properties of interface prototype object assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController interface: operation error(optional any) assert_own_property: self does not have own property "WritableStreamDefaultController" expected property "WritableStreamDefaultController" missing
FAIL WritableStreamDefaultController must be primary interface of self.writableStreamDefaultController assert_equals: wrong typeof object expected "object" but got "function"
FAIL Stringification of self.writableStreamDefaultController assert_equals: wrong typeof object expected "object" but got "function"
FAIL WritableStreamDefaultController interface: self.writableStreamDefaultController must inherit property "error(optional any)" with the proper type assert_equals: wrong typeof object expected "object" but got "function"
FAIL WritableStreamDefaultController interface: calling error(optional any) on self.writableStreamDefaultController with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "function"
FAIL TransformStream interface: existence and properties of interface object assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface object length assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface object name assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface: existence and properties of interface prototype object assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface: attribute readable assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream interface: attribute writable assert_own_property: self does not have own property "TransformStream" expected property "TransformStream" missing
FAIL TransformStream must be primary interface of new TransformStream() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TransformStream"
FAIL Stringification of new TransformStream() assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TransformStream"
FAIL TransformStream interface: new TransformStream() must inherit property "readable" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TransformStream"
FAIL TransformStream interface: new TransformStream() must inherit property "writable" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: TransformStream"
FAIL TransformStreamDefaultController interface: existence and properties of interface object assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface object length assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface object name assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: existence and properties of interface prototype object assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: attribute desiredSize assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: operation enqueue(optional any) assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: operation error(optional any) assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController interface: operation terminate() assert_own_property: self does not have own property "TransformStreamDefaultController" expected property "TransformStreamDefaultController" missing
FAIL TransformStreamDefaultController must be primary interface of self.transformStreamDefaultController assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of self.transformStreamDefaultController assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL TransformStreamDefaultController interface: self.transformStreamDefaultController must inherit property "desiredSize" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL TransformStreamDefaultController interface: self.transformStreamDefaultController must inherit property "enqueue(optional any)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL TransformStreamDefaultController interface: calling enqueue(optional any) on self.transformStreamDefaultController with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL TransformStreamDefaultController interface: self.transformStreamDefaultController must inherit property "error(optional any)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL TransformStreamDefaultController interface: calling error(optional any) on self.transformStreamDefaultController with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL TransformStreamDefaultController interface: self.transformStreamDefaultController must inherit property "terminate()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
PASS ByteLengthQueuingStrategy interface: existence and properties of interface object 
FAIL ByteLengthQueuingStrategy interface object length assert_equals: wrong value for ByteLengthQueuingStrategy.length expected 1 but got 0
PASS ByteLengthQueuingStrategy interface object name 
PASS ByteLengthQueuingStrategy interface: existence and properties of interface prototype object 
PASS ByteLengthQueuingStrategy interface: existence and properties of interface prototype object's "constructor" property 
PASS ByteLengthQueuingStrategy interface: existence and properties of interface prototype object's @@unscopables property 
FAIL ByteLengthQueuingStrategy interface: attribute highWaterMark assert_true: The prototype object must have a property "highWaterMark" expected true got false
FAIL ByteLengthQueuingStrategy interface: attribute size assert_throws_js: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw
PASS ByteLengthQueuingStrategy must be primary interface of new ByteLengthQueuingStrategy({ highWaterMark: 5 }) 
PASS Stringification of new ByteLengthQueuingStrategy({ highWaterMark: 5 }) 
FAIL ByteLengthQueuingStrategy interface: new ByteLengthQueuingStrategy({ highWaterMark: 5 }) must inherit property "highWaterMark" with the proper type assert_inherits: property "highWaterMark" found on object expected in prototype chain
PASS ByteLengthQueuingStrategy interface: new ByteLengthQueuingStrategy({ highWaterMark: 5 }) must inherit property "size" with the proper type 
PASS CountQueuingStrategy interface: existence and properties of interface object 
FAIL CountQueuingStrategy interface object length assert_equals: wrong value for CountQueuingStrategy.length expected 1 but got 0
PASS CountQueuingStrategy interface object name 
PASS CountQueuingStrategy interface: existence and properties of interface prototype object 
PASS CountQueuingStrategy interface: existence and properties of interface prototype object's "constructor" property 
PASS CountQueuingStrategy interface: existence and properties of interface prototype object's @@unscopables property 
FAIL CountQueuingStrategy interface: attribute highWaterMark assert_true: The prototype object must have a property "highWaterMark" expected true got false
FAIL CountQueuingStrategy interface: attribute size assert_throws_js: getting property on prototype object must throw TypeError function "function () {
    [native code]
}" did not throw
PASS CountQueuingStrategy must be primary interface of new CountQueuingStrategy({ highWaterMark: 5 }) 
PASS Stringification of new CountQueuingStrategy({ highWaterMark: 5 }) 
FAIL CountQueuingStrategy interface: new CountQueuingStrategy({ highWaterMark: 5 }) must inherit property "highWaterMark" with the proper type assert_inherits: property "highWaterMark" found on object expected in prototype chain
PASS CountQueuingStrategy interface: new CountQueuingStrategy({ highWaterMark: 5 }) must inherit property "size" with the proper type 
